Move CarouselSpot extension into separate files
